Stars, Reviews, Genres
added star ratings in a more visual form. made an input field for users to enter their name, rate the book, and leave a comment. made the genre tabs look more similar to the home page while only displaying the appropriate books. added the book covers to the book details as well as slightly adjusting the styling